refactor: move functions in separate files
